+++ src/bookmark_reader/target_writer.rs
@@ -16,6 +16,8 @@ where
 
         self.write_all(&json).map_err(BogrepError::WriteFile)?;
 
+        self.flush().map_err(BogrepError::FlushFile)?;
+
         // Rewind after writing.
         self.rewind().map_err(BogrepError::RewindFile)?;