+++ src/settings.rs
@@ -95,8 +95,8 @@ pub struct Settings {
     pub cache_mode: CacheMode,
     /// The maximal number of concurrent requests.
     ///
-    /// On macOS, no more than 100 open files are supported which is why 100 is
-    /// also used as a default for the number of concurrent requests.
+    /// On macOS, no more than 100 file descriptors are supported which is why
+    /// 100 is also used as a default for the number of concurrent requests.
     pub max_concurrent_requests: usize,
     /// The request timeout in milliseconds.
     pub request_timeout: u64,
@@ -169,6 +169,8 @@ impl Settings {
                 settings_path.display()
             ))?;
             settings_file.write_all(&settings_json)?;
+            settings_file.flush()?;
+
             Ok(settings)
         }
     }