+++ src/config.rs
@@ -70,10 +70,10 @@ impl Config {
 
         let settings = Settings::init(&settings_path)?;
 
-        // `max_concurrent_requests` is bounded by how fast we can write the
+        //  `max_concurrent_requests` is bounded by how fast we can write the
         // content, i.e. by the maximum number of open file descriptors.
         #[cfg(not(any(target_os = "windows")))]
-        set_file_descriptor_limit(settings.max_concurrent_requests as u64)?;
+        set_file_descriptor_limit(settings.max_concurrent_requests as u64 * 2)?;
 
         if !target_bookmark_path.exists() {
             debug!(