+++ src/settings.rs
@@ -13,7 +13,10 @@ use std::{
     path::Path,
 };
 
-/// The default `Settungs::max_concurrent_requests`.
+/// The default for `Settungs::max_open_files`.
+const MAX_OPEN_FILES_DEFAULT: u64 = 100;
+
+/// The default for `Settungs::max_concurrent_requests`.
 const MAX_CONCURRENT_REQUESTS_DEFAULT: usize = 100;
 
 /// The default for `Settings::request_timeout`.
@@ -41,6 +44,9 @@ pub struct SettingsArgs {
     pub underlying_urls: Vec<String>,
     /// The file extension used to cache websites.
     pub cache_mode: Option<CacheMode>,
+    /// The maximal number of open files used to write the fetched content to
+    /// disk.
+    pub max_open_files: Option<u64>,
     /// The maximal number of concurrent requests.
     pub max_concurrent_requests: Option<usize>,
     /// The request timeout in milliseconds.
@@ -61,6 +67,7 @@ impl SettingsArgs {
         ignored_urls: Vec<String>,
         underlying_urls: Vec<String>,
         cache_mode: Option<CacheMode>,
+        max_open_files: Option<u64>,
         max_concurrent_requests: Option<usize>,
         request_timeout: Option<u64>,
         request_throttling: Option<u64>,
@@ -72,6 +79,7 @@ impl SettingsArgs {
             ignored_urls,
             underlying_urls,
             cache_mode,
+            max_open_files,
             max_concurrent_requests,
             request_timeout,
             request_throttling,
@@ -94,9 +102,13 @@ pub struct Settings {
     pub underlying_urls: Vec<String>,
     /// The file extension used to cache websites.
     pub cache_mode: CacheMode,
-    /// The maximal number of concurrent requests.
+    /// The maximal number of open files used to write the fetched content to
+    /// disk.
+    pub max_open_files: u64,
+    /// The maximal number of concurrent requests used to fetch the content from
+    /// the bookmarked websites.
     ///
-    /// On macOS, no more than 100 file descriptors are supported which is why
+    /// On macOS, no more than 256 file descriptors are supported which is why
     /// 100 is also used as a default for the number of concurrent requests.
     pub max_concurrent_requests: usize,
     /// The request timeout in milliseconds.
@@ -107,18 +119,18 @@ pub struct Settings {
     /// the client.
     ///
     /// As the maximum number of open file descriptors (i.e. open files +
-    /// network sockets) are limited, `max_idle_connections_per_host` should be
-    /// configured to 1.
+    /// network sockets) are limited by the operating system,
+    /// `max_idle_connections_per_host` should be configured to 1.
     ///
-    /// Fix "Too many open files" and DNS errors (rate limit for DNS server) by
-    /// choosing a sensible value for `max_idle_connections_per_host` and
-    /// `idle_connections_timeout`.
+    /// Choose sensible values for `max_idle_connections_per_host` and
+    /// `idle_connections_timeout` to fix "Too many open files" and DNS errors
+    /// (rate limit for DNS server).
     pub max_idle_connections_per_host: usize,
     /// The timeout for idle connections to be kept alive in milliseconds.
     ///
-    /// Fix "Too many open files" and DNS errors (rate limit for DNS server) by
-    /// choosing a sensible value for `max_idle_connections_per_host` and
-    /// `idle_connections_timeout`.
+    /// Choose sensible values for `max_idle_connections_per_host` and
+    /// `idle_connections_timeout` to fix "Too many open files" and DNS errors
+    /// (rate limit for DNS server).
     pub idle_connections_timeout: u64,
 }
 
@@ -129,6 +141,7 @@ impl Default for Settings {
             ignored_urls: Vec::new(),
             underlying_urls: Vec::new(),
             cache_mode: CacheMode::default(),
+            max_open_files: MAX_OPEN_FILES_DEFAULT,
             max_concurrent_requests: MAX_CONCURRENT_REQUESTS_DEFAULT,
             request_timeout: REQUEST_TIMEOUT_DEFAULT,
             request_throttling: REQUEST_THROTTLING_DEFAULT,
@@ -145,6 +158,7 @@ impl Settings {
         ignored_urls: Vec<String>,
         underlying_urls: Vec<String>,
         cache_mode: CacheMode,
+        max_open_files: u64,
         max_concurrent_requests: usize,
         request_timeout: u64,
         request_throttling: u64,
@@ -156,6 +170,7 @@ impl Settings {
             ignored_urls,
             underlying_urls,
             cache_mode,
+            max_open_files,
             max_concurrent_requests,
             request_timeout,
             request_throttling,
@@ -243,6 +258,11 @@ impl Settings {
         }
     }
 
+    pub fn set_max_open_files(&mut self, max_open_files: u64) {
+        debug!("Set `max_open_files` to {max_open_files}");
+        self.max_open_files = max_open_files;
+    }
+
     pub fn set_request_timeout(&mut self, request_timeout: u64) {
         debug!("Set `request_timeout` to {request_timeout}");
         self.request_timeout = request_timeout;
@@ -254,7 +274,7 @@ impl Settings {
     }
 
     pub fn set_max_concurrent_requests(&mut self, max_concurrent_requests: usize) {
-        debug!("Set `max_concurrent_requests` timeout to {max_concurrent_requests}");
+        debug!("Set `max_concurrent_requests` to {max_concurrent_requests}");
         self.max_concurrent_requests = max_concurrent_requests;
     }